requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
PyMySQL==1.1.0
sqlalchemy==2.0.36
python-dotenv==1.0.0
starlette==0.27.0
alembic==1.13.3
cryptography